Updated methods and JUnit Tests
Program seems to be working for Tasks 1 and 2 now

+++ 594 Group Project/src/edu/upenn/cit594/data/Property.java
@@ -2,12 +2,12 @@ package edu.upenn.cit594.data;
 
 public class Property {
 
-	Integer marketValue; 
-	Integer livableArea;
+	double marketValue; 
+	double livableArea;
 	String zipCode;
 	
 	
-	public Property(Integer marketValue, Integer livableArea, String zipCode) {
+	public Property(double marketValue, double livableArea, String zipCode) {
 		super();
 		this.marketValue = marketValue;
 		this.livableArea = livableArea;
